+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Unreleased
 
+- [...]
+
+# v41.7.0 (16/11/2020)
+
 - **[UPDATE]** New hooks for `MediaSizeProvider` and add a `mediaSizeForTestsOnly` override for test setup.
 - **[FIX]** Make `OnChangeParameters` part of the public API.
 - **[NEW]** Add `SupportIcon`
-
 
 # v41.6.3 (09/11/2020)